# krt: optimize lookups in Join