```
cpu: AMD Ryzen 7 PRO 7840U w/ Radeon 780M Graphics
                                                           │  baseline2  │                join2                │
                                                           │   sec/op    │   sec/op     vs base                │
AddressFullGeneration/serviceentry-workloadentry-16          1.377m ± 4%   1.115m ± 3%  -19.05% (p=0.000 n=10)
AddressIncrementalGeneration/serviceentry-workloadentry-16   1.410µ ± 7%   1.411µ ± 3%        ~ (p=0.869 n=10)
geomean                                                      44.06µ        39.66µ        -9.99%

                                                           │   baseline2   │                 join2                  │
                                                           │     B/op      │     B/op      vs base                  │
AddressFullGeneration/serviceentry-workloadentry-16          1254.2Ki ± 0%   995.8Ki ± 0%  -20.61% (p=0.000 n=10)
AddressIncrementalGeneration/serviceentry-workloadentry-16    1.250Ki ± 0%   1.250Ki ± 0%        ~ (p=1.000 n=10) ¹
geomean                                                       39.59Ki        35.28Ki       -10.90%
¹ all samples are equal

                                                           │  baseline2  │                 join2                 │
                                                           │  allocs/op  │  allocs/op   vs base                  │
AddressFullGeneration/serviceentry-workloadentry-16          9.182k ± 0%   7.124k ± 0%  -22.41% (p=0.000 n=10)
AddressIncrementalGeneration/serviceentry-workloadentry-16    18.00 ± 0%    18.00 ± 0%        ~ (p=1.000 n=10) ¹
geomean                                                       406.5         358.1       -11.92%
¹ all samples are equal
```