@istio-testing 2025-01-03
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
gencheck_istio | f566aac2c59abaa9a035efc4bf924c714442aa22 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/54551/gencheck_istio/1875313371633422336) | true | `/test gencheck`
lint_istio | f566aac2c59abaa9a035efc4bf924c714442aa22 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/54551/lint_istio/1875313380026224640) | true | `/test lint`
integ-ambient_istio | f566aac2c59abaa9a035efc4bf924c714442aa22 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/54551/integ-ambient_istio/1875313371683753984) | true | `/test integ-ambient`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes-sigs/prow](https://github.com/kubernetes-sigs/prow/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->